Break facet fields out from browser into a separate function
